requirements.txt
opencv-python>=4.8.0
openai-whisper>=20231117
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
